requirements.txt
newspaper3k
requests
boto3
bs4
pandas
nltk
pickleshare

